associateTokens to command
